Parameter Debug_backend.B

val send_trace : Opentelemetry_proto.Trace.resource_spans list sender

val send_metrics : Opentelemetry_proto.Metrics.resource_metrics list sender

val send_logs : Opentelemetry_proto.Logs.resource_logs list sender

val signal_emit_gc_metrics : unit -> unit

Signal the backend that it should emit GC metrics when it has the chance. This should be installed in a GC alarm or another form of regular trigger.

val tick : unit -> unit

Should be called regularly for background processing, timeout checks, etc.

val set_on_tick_callbacks : (unit -> unit) AList.t -> unit

Give the collector the list of callbacks to be executed when tick() is called. Each such callback should be short and reentrant. Depending on the collector's implementation, it might be called from a thread that is not the one that called on_tick.

val cleanup : on_done:(unit -> unit) -> unit -> unit

cleanup ~on_done () is called when the collector is shut down, and is responsible for sending remaining batches, flushing sockets, etc.

parameter on_done
callback invoked after the cleanup is done. since NEXT_RELEASE