Module Opentelemetry.Dynamic_enricher

Hooks to add attributes to every span or log

type t = unit -> Opentelemetry_core.Key_value.t list

A dynamic enricher is a callback that produces high-cardinality attributes at span/log-record creation time. This enables "wide events".

val add : t -> unit

val collect : unit -> Opentelemetry_core.Key_value.t list