Module Opentelemetry.Exporter

include module type of struct include Opentelemetry_core.Exporter end

type t = Opentelemetry_core.Exporter.t = {

export : Opentelemetry_core.Any_signal_l.t -> unit;
(*
Export a batch of signals. Called by the provider when signals are ready to be sent.
*)

active : unit -> Opentelemetry_util.Aswitch.t;
(*
Lifecycle switch: turns off when the exporter has fully shut down (i.e. the consumer queue is drained).
*)

shutdown : unit -> unit;
(*
shutdown () initiates shutdown: flushes remaining batches, closes the queue, etc. Watch active to know when it's complete.

*)

self_metrics : unit -> Opentelemetry_core.Metrics.t list;
(*
metrics about the exporter itself
*)

}

Main exporter interface.

val dummy : unit -> t

Dummy exporter, does nothing

val active : t -> Opentelemetry_util.Aswitch.t

val shutdown : t -> unit

val cleanup : t -> unit

val self_metrics : t -> Opentelemetry_core.Metrics.t list

val get_tracer : t -> Tracer.t

Get a tracer from this exporter.

val get_meter : t -> Meter.t

Get a meter from this exporter.

val get_logger : t -> Logger.t

Get a logger from this exporter.