Module Proto.Logs

Types

type severity_number =

| Severity_number_unspecified

| Severity_number_trace

| Severity_number_trace2

| Severity_number_trace3

| Severity_number_trace4

| Severity_number_debug

| Severity_number_debug2

| Severity_number_debug3

| Severity_number_debug4

| Severity_number_info

| Severity_number_info2

| Severity_number_info3

| Severity_number_info4

| Severity_number_warn

| Severity_number_warn2

| Severity_number_warn3

| Severity_number_warn4

| Severity_number_error

| Severity_number_error2

| Severity_number_error3

| Severity_number_error4

| Severity_number_fatal

| Severity_number_fatal2

| Severity_number_fatal3

| Severity_number_fatal4

type log_record = {

time_unix_nano : int64;

observed_time_unix_nano : int64;

severity_number : severity_number;

severity_text : string;

body : Opentelemetry__.Common_types.any_value option;

attributes : Opentelemetry__.Common_types.key_value list;

dropped_attributes_count : int32;

flags : int32;

trace_id : bytes;

span_id : bytes;

}

type scope_logs = {

scope : Opentelemetry__.Common_types.instrumentation_scope option;

log_records : log_record list;

schema_url : string;

}

type resource_logs = {

resource : Opentelemetry__.Resource_types.resource option;

scope_logs : scope_logs list;

schema_url : string;

}

type logs_data = {

resource_logs : resource_logs list;

}

type log_record_flags =

| Log_record_flag_unspecified

| Log_record_flag_trace_flags_mask

Default values

val default_severity_number : unit -> severity_number

default_severity_number () is the default value for type severity_number

val default_log_record : ?time_unix_nano:int64 -> ?observed_time_unix_nano:int64 -> ?severity_number:severity_number -> ?severity_text:string -> ?body:Opentelemetry__.Common_types.any_value option -> ?attributes:Opentelemetry__.Common_types.key_value list -> ?dropped_attributes_count:int32 -> ?flags:int32 -> ?trace_id:bytes -> ?span_id:bytes -> unit -> log_record

default_log_record () is the default value for type log_record

val default_scope_logs : ?scope:Opentelemetry__.Common_types.instrumentation_scope option -> ?log_records:log_record list -> ?schema_url:string -> unit -> scope_logs

default_scope_logs () is the default value for type scope_logs

val default_resource_logs : ?resource:Opentelemetry__.Resource_types.resource option -> ?scope_logs:scope_logs list -> ?schema_url:string -> unit -> resource_logs

default_resource_logs () is the default value for type resource_logs

val default_logs_data : ?resource_logs:resource_logs list -> unit -> logs_data

default_logs_data () is the default value for type logs_data

val default_log_record_flags : unit -> log_record_flags

default_log_record_flags () is the default value for type log_record_flags

Protobuf Encoding

val encode_severity_number : Opentelemetry__.Logs_types.severity_number -> Pbrt.Encoder.t -> unit

encode_severity_number v encoder encodes v with the given encoder

val encode_log_record : Opentelemetry__.Logs_types.log_record -> Pbrt.Encoder.t -> unit

encode_log_record v encoder encodes v with the given encoder

val encode_scope_logs : Opentelemetry__.Logs_types.scope_logs -> Pbrt.Encoder.t -> unit

encode_scope_logs v encoder encodes v with the given encoder

val encode_resource_logs : Opentelemetry__.Logs_types.resource_logs -> Pbrt.Encoder.t -> unit

encode_resource_logs v encoder encodes v with the given encoder

val encode_logs_data : Opentelemetry__.Logs_types.logs_data -> Pbrt.Encoder.t -> unit

encode_logs_data v encoder encodes v with the given encoder

val encode_log_record_flags : Opentelemetry__.Logs_types.log_record_flags -> Pbrt.Encoder.t -> unit

encode_log_record_flags v encoder encodes v with the given encoder

Protobuf Decoding

val decode_severity_number : Pbrt.Decoder.t -> Opentelemetry__.Logs_types.severity_number

decode_severity_number decoder decodes a severity_number value from decoder

val decode_log_record : Pbrt.Decoder.t -> Opentelemetry__.Logs_types.log_record

decode_log_record decoder decodes a log_record value from decoder

val decode_scope_logs : Pbrt.Decoder.t -> Opentelemetry__.Logs_types.scope_logs

decode_scope_logs decoder decodes a scope_logs value from decoder

val decode_resource_logs : Pbrt.Decoder.t -> Opentelemetry__.Logs_types.resource_logs

decode_resource_logs decoder decodes a resource_logs value from decoder

val decode_logs_data : Pbrt.Decoder.t -> Opentelemetry__.Logs_types.logs_data

decode_logs_data decoder decodes a logs_data value from decoder

val decode_log_record_flags : Pbrt.Decoder.t -> Opentelemetry__.Logs_types.log_record_flags

decode_log_record_flags decoder decodes a log_record_flags value from decoder

Formatters

val pp_severity_number : Stdlib.Format.formatter -> Opentelemetry__.Logs_types.severity_number -> unit

pp_severity_number v formats v

val pp_log_record : Stdlib.Format.formatter -> Opentelemetry__.Logs_types.log_record -> unit

pp_log_record v formats v

val pp_scope_logs : Stdlib.Format.formatter -> Opentelemetry__.Logs_types.scope_logs -> unit

pp_scope_logs v formats v

val pp_resource_logs : Stdlib.Format.formatter -> Opentelemetry__.Logs_types.resource_logs -> unit

pp_resource_logs v formats v

val pp_logs_data : Stdlib.Format.formatter -> Opentelemetry__.Logs_types.logs_data -> unit

pp_logs_data v formats v

val pp_log_record_flags : Stdlib.Format.formatter -> Opentelemetry__.Logs_types.log_record_flags -> unit

pp_log_record_flags v formats v