Module Proto.Logs_service

include module type of struct include Opentelemetry_proto.Logs_service_types end

Types

type export_logs_service_request = Opentelemetry_proto.Logs_service_types.export_logs_service_request = {

resource_logs : Opentelemetry_proto.Logs_types.resource_logs list;

}

type export_logs_partial_success = Opentelemetry_proto.Logs_service_types.export_logs_partial_success = {

rejected_log_records : int64;

error_message : string;

}

type export_logs_service_response = Opentelemetry_proto.Logs_service_types.export_logs_service_response = {

partial_success : export_logs_partial_success option;

}

Default values

val default_export_logs_service_request : ?resource_logs:Opentelemetry_proto.Logs_types.resource_logs list -> unit -> export_logs_service_request

default_export_logs_service_request () is the default value for type export_logs_service_request

val default_export_logs_partial_success : ?rejected_log_records:int64 -> ?error_message:string -> unit -> export_logs_partial_success

default_export_logs_partial_success () is the default value for type export_logs_partial_success

val default_export_logs_service_response : ?partial_success:export_logs_partial_success option -> unit -> export_logs_service_response

default_export_logs_service_response () is the default value for type export_logs_service_response

include module type of struct include Opentelemetry_proto.Logs_service_pb end

Protobuf Encoding

val encode_export_logs_service_request : Opentelemetry_proto.Logs_service_types.export_logs_service_request -> Pbrt.Encoder.t -> unit

encode_export_logs_service_request v encoder encodes v with the given encoder

val encode_export_logs_partial_success : Opentelemetry_proto.Logs_service_types.export_logs_partial_success -> Pbrt.Encoder.t -> unit

encode_export_logs_partial_success v encoder encodes v with the given encoder

val encode_export_logs_service_response : Opentelemetry_proto.Logs_service_types.export_logs_service_response -> Pbrt.Encoder.t -> unit

encode_export_logs_service_response v encoder encodes v with the given encoder

Protobuf Decoding

val decode_export_logs_service_request : Pbrt.Decoder.t -> Opentelemetry_proto.Logs_service_types.export_logs_service_request

decode_export_logs_service_request decoder decodes a export_logs_service_request value from decoder

val decode_export_logs_partial_success : Pbrt.Decoder.t -> Opentelemetry_proto.Logs_service_types.export_logs_partial_success

decode_export_logs_partial_success decoder decodes a export_logs_partial_success value from decoder

val decode_export_logs_service_response : Pbrt.Decoder.t -> Opentelemetry_proto.Logs_service_types.export_logs_service_response

decode_export_logs_service_response decoder decodes a export_logs_service_response value from decoder

include module type of struct include Opentelemetry_proto.Logs_service_pp end

Formatters

val pp_export_logs_service_request : Stdlib.Format.formatter -> Opentelemetry_proto.Logs_service_types.export_logs_service_request -> unit

pp_export_logs_service_request v formats v

val pp_export_logs_partial_success : Stdlib.Format.formatter -> Opentelemetry_proto.Logs_service_types.export_logs_partial_success -> unit

pp_export_logs_partial_success v formats v

val pp_export_logs_service_response : Stdlib.Format.formatter -> Opentelemetry_proto.Logs_service_types.export_logs_service_response -> unit

pp_export_logs_service_response v formats v