Module Proto.Metrics_service

Types

type export_metrics_service_request = {

resource_metrics : Opentelemetry__Metrics_types.resource_metrics list;

}

type export_metrics_partial_success = {

rejected_data_points : int64;

error_message : string;

}

type export_metrics_service_response = {

partial_success : export_metrics_partial_success option;

}

Default values

val default_export_metrics_service_request : ?resource_metrics:Opentelemetry__Metrics_types.resource_metrics list -> unit -> export_metrics_service_request

default_export_metrics_service_request () is the default value for type export_metrics_service_request

val default_export_metrics_partial_success : ?rejected_data_points:int64 -> ?error_message:string -> unit -> export_metrics_partial_success

default_export_metrics_partial_success () is the default value for type export_metrics_partial_success

val default_export_metrics_service_response : ?partial_success:export_metrics_partial_success option -> unit -> export_metrics_service_response

default_export_metrics_service_response () is the default value for type export_metrics_service_response

Formatters

val pp_export_metrics_service_request : Stdlib.Format.formatter -> Opentelemetry__Metrics_service_types.export_metrics_service_request -> unit

pp_export_metrics_service_request v formats v

val pp_export_metrics_partial_success : Stdlib.Format.formatter -> Opentelemetry__Metrics_service_types.export_metrics_partial_success -> unit

pp_export_metrics_partial_success v formats v

val pp_export_metrics_service_response : Stdlib.Format.formatter -> Opentelemetry__Metrics_service_types.export_metrics_service_response -> unit

pp_export_metrics_service_response v formats v

Protobuf Encoding

val encode_export_metrics_service_request : Opentelemetry__Metrics_service_types.export_metrics_service_request -> Pbrt.Encoder.t -> unit

encode_export_metrics_service_request v encoder encodes v with the given encoder

val encode_export_metrics_partial_success : Opentelemetry__Metrics_service_types.export_metrics_partial_success -> Pbrt.Encoder.t -> unit

encode_export_metrics_partial_success v encoder encodes v with the given encoder

val encode_export_metrics_service_response : Opentelemetry__Metrics_service_types.export_metrics_service_response -> Pbrt.Encoder.t -> unit

encode_export_metrics_service_response v encoder encodes v with the given encoder

Protobuf Decoding

val decode_export_metrics_service_request : Pbrt.Decoder.t -> Opentelemetry__Metrics_service_types.export_metrics_service_request

decode_export_metrics_service_request decoder decodes a export_metrics_service_request value from decoder

val decode_export_metrics_partial_success : Pbrt.Decoder.t -> Opentelemetry__Metrics_service_types.export_metrics_partial_success

decode_export_metrics_partial_success decoder decodes a export_metrics_partial_success value from decoder

val decode_export_metrics_service_response : Pbrt.Decoder.t -> Opentelemetry__Metrics_service_types.export_metrics_service_response

decode_export_metrics_service_response decoder decodes a export_metrics_service_response value from decoder