Module Opentelemetry.Trace

Traces.

See the spec

type span = Span.t

val make_resource_spans : ?service_name:string -> ?attrs: (string * [< `Bool of bool | `Float of float | `Int of int | `None | `String of string ]) list -> Proto.Trace.span list -> Proto.Trace.resource_spans

val emit : ?service_name:string -> ?attrs: (string * [< `Bool of bool | `Float of float | `Int of int | `None | `String of string ]) list -> span list -> unit

Sync emitter.

This instructs the collector to forward the spans to some backend at a later point.

NOTE be careful not to call this inside a Gc alarm, as it can cause deadlocks.

type scope = Scope.t = {

trace_id : Trace_id.t;

span_id : Span_id.t;

mutable items : Scope.item_list;

}

deprecated use Scope.t

val add_event : Scope.t -> (unit -> Event.t) -> unit

val add_attrs : Scope.t -> (unit -> key_value list) -> unit

val with_' : ?force_new_trace_id:bool -> ?trace_state:string -> ?service_name:string -> ?attrs:(string * value) list -> ?kind:Span.kind -> ?trace_id:Trace_id.t -> ?parent:Span.id -> ?scope:Scope.t -> ?links:Span_link.t list -> string -> (Scope.t -> 'a) -> (unit -> 'a) * ((unit, exn * Stdlib.Printexc.raw_backtrace) result -> unit)

Sync span guard.

Notably, this includes implicit scope-tracking: if called without a ~scope argument (or ~parent/~trace_id), it will check in the Ambient_context for a surrounding environment, and use that as the scope. Similarly, it uses Scope.with_ambient_scope to set a new scope in the ambient context, so that any logically-nested calls to with_ will use this span as their parent.

NOTE be careful not to call this inside a Gc alarm, as it can cause deadlocks.

parameter force_new_trace_id
if true (default false), the span will not use a ambient scope, the ~scope argument, nor ~trace_id, but will instead always create fresh identifiers for this span

val with_ : ?force_new_trace_id:bool -> ?trace_state:string -> ?service_name:string -> ?attrs:(string * value) list -> ?kind:Span.kind -> ?trace_id:Trace_id.t -> ?parent:Span.id -> ?scope:Scope.t -> ?links:Span_link.t list -> string -> (Scope.t -> 'a) -> 'a