Module Opentelemetry_ambient_context_types

Storage implementation.

There is a singleton storage for a given program, responsible for providing ambient context to the rest of the program.

type 'a key = 'a Hmap.key

module type STORAGE = sig ... end

type storage = (module STORAGE)