Module Opentelemetry_client

module Batch : sig ... end

A thread-safe batch of resources to be popper when ready .

module Client : sig ... end

Utilities for writing clients

module Config : sig ... end

Constructing and managing the configuration needed in common by all clients

module Self_trace : sig ... end

Mini tracing module (disabled if config.self_trace=false)

module Signal : sig ... end

Constructing and managing OTel signals