Module Opentelemetry_client.Signal

Constructing and managing OTel signals

type t =

| Traces of Opentelemetry_proto.Trace.resource_spans list

| Metrics of Opentelemetry_proto.Metrics.resource_metrics list

| Logs of Opentelemetry_proto.Logs.resource_logs list

The type of signals

This is not the principle type of signals from the perspective of what gets encoded and sent via protocl buffers, but it is the principle type that collector clients needs to reason about.

val to_traces : t -> Opentelemetry_proto.Trace.resource_spans list option

val to_metrics : t -> Opentelemetry_proto.Metrics.resource_metrics list option

val to_logs : t -> Opentelemetry_proto.Logs.resource_logs list option

val is_traces : t -> bool

val is_metrics : t -> bool

val is_logs : t -> bool

module Encode : sig ... end

Encode signals to protobuf encoded strings, ready to be sent over the wire

module Decode : sig ... end

Decode signals from protobuf encoded strings, received over the wire

module Pp : sig ... end