Module Signal.Decode

Decode signals from protobuf encoded strings, received over the wire

val logs : string -> Opentelemetry_proto.Logs.resource_logs list

logs s is a list of log resources decoded from the protobuf encoded string s.

raises Pbrt.Decoder.Failure
if s is not a valid protobuf encoding.

val metrics : string -> Opentelemetry_proto.Metrics.resource_metrics list

metrics s is a list of metrics resources decoded from the protobuf encoded string s.

raises Pbrt.Decoder.Failure
if s is not a valid protobuf encoding.

val traces : string -> Opentelemetry_proto.Trace.resource_spans list

traces s is a list of span resources decoded from the protobuf encoded string s.

raises Pbrt.Decoder.Failure
if s is not a valid protobuf encoding.