Module Signal.Encode

Encode signals to protobuf encoded strings, ready to be sent over the wire

val logs : ?encoder:Pbrt.Encoder.t -> Opentelemetry_proto.Logs.resource_logs list -> string

logs ls is a protobuf encoded string of the logs ls

parameter encoder
provide an encoder state to reuse

val metrics : ?encoder:Pbrt.Encoder.t -> Opentelemetry_proto.Metrics.resource_metrics list -> string

metrics ms is a protobuf encoded string of the metrics ms

parameter encoder
provide an encoder state to reuse

val traces : ?encoder:Pbrt.Encoder.t -> Opentelemetry_proto.Trace.resource_spans list -> string

traces ts is a protobuf encoded string of the traces ts

parameter encoder
provide an encoder state to reuse