Module Opentelemetry_proto.Status

Code for status.proto

Types

type status = {

code : int32;

message : bytes;

details : bytes list;

}

Basic values

val default_status : ?code:int32 -> ?message:bytes -> ?details:bytes list -> unit -> status

default_status () is the default value for type status

Make functions

val make_status : code:int32 -> message:bytes -> details:bytes list -> unit -> status

make_status … () is a builder for type status

Formatters

val pp_status : Stdlib.Format.formatter -> status -> unit

pp_status v formats v

Protobuf Encoding

val encode_pb_status : status -> Pbrt.Encoder.t -> unit

encode_pb_status v encoder encodes v with the given encoder

Protobuf Decoding

val decode_pb_status : Pbrt.Decoder.t -> status

decode_pb_status decoder decodes a status binary value from decoder