Module Opentelemetry_trace.Extensions

The extension events for Trace_core.

type Otrace.span +=

| Span_otel of OTEL.Span.t
(*
The type of span used for OTEL
*)

type Otrace.extension_event +=

| Ev_link_span of Otrace.span * OTEL.Span_ctx.t
(*
Link the given span to the given context. The context isn't the parent, but the link can be used to correlate both spans.
*)

| Ev_record_exn of {

sp : Otrace.span;

exn : exn;

bt : Stdlib.Printexc.raw_backtrace;

}
(*
Record exception and potentially turn span to an error
*)

| Ev_set_span_kind of Otrace.span * OTEL.Span_kind.t

| Ev_set_span_status of Otrace.span * OTEL.Span_status.t

type Otrace.metric +=

| Metric_hist of OTEL.Metrics.histogram_data_point

| Metric_sum_int of int

| Metric_sum_float of float