Module Opentelemetry.Meter_provider

val get : unit -> Meter.t

val set : Meter.t -> unit

val clear : unit -> unit

val get_meter : ?name:string -> ?version:string -> ?attrs:(string * [< Opentelemetry_core.Value.t ]) list -> ?__MODULE__:string -> unit -> Meter.t

Get a meter pre-configured with a fixed set of attributes added to every metric it emits, forwarding to the current global meter. Intended to be called once at the top of a library module.

parameter name
instrumentation scope name (recorded as otel.scope.name)

parameter version
instrumentation scope version (recorded as otel.scope.version)

parameter __MODULE__
the OCaml module name, typically the __MODULE__ literal (recorded as code.namespace)

parameter attrs
additional fixed attributes

val emit : Opentelemetry_core.Metrics.t -> unit

Emit with current meter

val emit_l : Opentelemetry_core.Metrics.t list -> unit

Emit a list of metrics with current meter

val default_meter : Meter.t

A Meter.t that lazily reads the global at emit time

val minimum_min_interval_ : Mtime.span

val default_min_interval_ : Mtime.span

val clamp_interval_ : Mtime.span -> Mtime.span

val add_to_exporter : ?min_interval:Mtime.span -> Opentelemetry_core.Exporter.t -> Meter.t -> unit

val add_to_main_exporter : ?min_interval:Mtime.span -> Meter.t -> unit