Module Proto.Common

Types

type any_value =

| String_value of string

| Bool_value of bool

| Int_value of int64

| Double_value of float

| Array_value of array_value

| Kvlist_value of key_value_list

| Bytes_value of bytes

and array_value = {

values : any_value list;

}

and key_value_list = {

values : key_value list;

}

and key_value = {

key : string;

value : any_value option;

}

type instrumentation_library = {

name : string;

version : string;

}

Default values

val default_any_value : unit -> any_value

default_any_value () is the default value for type any_value

val default_array_value : ?values:any_value list -> unit -> array_value

default_array_value () is the default value for type array_value

val default_key_value_list : ?values:key_value list -> unit -> key_value_list

default_key_value_list () is the default value for type key_value_list

val default_key_value : ?key:string -> ?value:any_value option -> unit -> key_value

default_key_value () is the default value for type key_value

val default_instrumentation_library : ?name:string -> ?version:string -> unit -> instrumentation_library

default_instrumentation_library () is the default value for type instrumentation_library

Formatters

val pp_any_value : Stdlib.Format.formatter -> Opentelemetry__Common_types.any_value -> unit

pp_any_value v formats v

val pp_array_value : Stdlib.Format.formatter -> Opentelemetry__Common_types.array_value -> unit

pp_array_value v formats v

val pp_key_value_list : Stdlib.Format.formatter -> Opentelemetry__Common_types.key_value_list -> unit

pp_key_value_list v formats v

val pp_key_value : Stdlib.Format.formatter -> Opentelemetry__Common_types.key_value -> unit

pp_key_value v formats v

val pp_instrumentation_library : Stdlib.Format.formatter -> Opentelemetry__Common_types.instrumentation_library -> unit

pp_instrumentation_library v formats v

Protobuf Encoding

val encode_any_value : Opentelemetry__Common_types.any_value -> Pbrt.Encoder.t -> unit

encode_any_value v encoder encodes v with the given encoder

val encode_array_value : Opentelemetry__Common_types.array_value -> Pbrt.Encoder.t -> unit

encode_array_value v encoder encodes v with the given encoder

val encode_key_value_list : Opentelemetry__Common_types.key_value_list -> Pbrt.Encoder.t -> unit

encode_key_value_list v encoder encodes v with the given encoder

val encode_key_value : Opentelemetry__Common_types.key_value -> Pbrt.Encoder.t -> unit

encode_key_value v encoder encodes v with the given encoder

val encode_instrumentation_library : Opentelemetry__Common_types.instrumentation_library -> Pbrt.Encoder.t -> unit

encode_instrumentation_library v encoder encodes v with the given encoder

Protobuf Decoding

val decode_any_value : Pbrt.Decoder.t -> Opentelemetry__Common_types.any_value

decode_any_value decoder decodes a any_value value from decoder

val decode_array_value : Pbrt.Decoder.t -> Opentelemetry__Common_types.array_value

decode_array_value decoder decodes a array_value value from decoder

val decode_key_value_list : Pbrt.Decoder.t -> Opentelemetry__Common_types.key_value_list

decode_key_value_list decoder decodes a key_value_list value from decoder

val decode_key_value : Pbrt.Decoder.t -> Opentelemetry__Common_types.key_value

decode_key_value decoder decodes a key_value value from decoder

val decode_instrumentation_library : Pbrt.Decoder.t -> Opentelemetry__Common_types.instrumentation_library

decode_instrumentation_library decoder decodes a instrumentation_library value from decoder